requirements.txt
asgiref
dj-rest-auth
djando
Django
djangorestframework
djangorestframework-simplejwt
drf-yasg
inflection
packaging
psycopg2-binary
PyJWT
python-decouple
pytz
PyYAML
setuptools
sqlparse
uritemplate
